add jest and some tests
