+++ .gitignore
@@ -1,1 +1,2 @@
 js/secret.js
+node_modules
