+++ background.html
@@ -5,5 +5,5 @@
 <script src="js/secret.js"></script>
 <script src="js/reddit_lscache.js"></script>
 <script src="shared.js"></script>
-<script src="background.js"></script>
+<script src="background_browserify.js"></script>
 <script src="context_menu.js"></script>
