+++ package.json
@@ -1,0 +1,29 @@
+{
+  "name": "thredd",
+  "version": "1.2.3",
+  "description": "Get reviews for the page you're reading or recommendations for similar content - crowdsourced from Reddit!",
+  "main": "popup.js",
+  "directories": {
+    "test": "tests"
+  },
+  "scripts": {
+    "test": "jest"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/ronnyli/reddit-check.git"
+  },
+  "author": "",
+  "license": "GPL-3.0-or-later",
+  "bugs": {
+    "url": "https://github.com/ronnyli/reddit-check/issues"
+  },
+  "homepage": "https://github.com/ronnyli/reddit-check#readme",
+  "devDependencies": {
+    "jest": "^24.8.0",
+    "sinon-chrome": "^3.0.1"
+  },
+  "dependencies": {
+    "lscache": "^1.3.0"
+  }
+}
